Created URLRequest and URLSession (moved methods from OAuth2Service class).
