+++ ImageFeed/OAuth2/OAuth2Service.swift
@@ -103,53 +103,3 @@ extension OAuth2Service {
             baseURL: Constants.baseURL)
     }
 }
-
-// MARK: - HTTP Request
-
-extension URLRequest {
-    static func makeHTTPRequest(path: String,
-                                httpMethod: String,
-                                baseURL: URL = Constants.defaultbaseURL) -> URLRequest {
-        var request = URLRequest(url: URL(string: path, relativeTo: baseURL)!)
-        request.httpMethod = httpMethod
-        return request
-    }
-}
-
-// MARK: - Network Connection
-
-enum NetworkError: Error {
-    case httpStatusCode(Int)
-    case urlRequestError(Error)
-    case urlSessionError
-}
-
-extension URLSession {
-    func data(for request: URLRequest,
-              completion: @escaping (Result<Data, Error>) -> Void)
-    -> URLSessionTask {
-        let fulfillCompletion: (Result<Data, Error>) -> Void = { result in
-            DispatchQueue.main.async {
-                completion(result)
-            }
-        }
-
-        let task = dataTask(with: request, completionHandler: { data, response, error in
-            if let data = data,
-                let response = response,
-                let statusCode = (response as? HTTPURLResponse)?.statusCode
-            {
-                if 200 ..< 300 ~= statusCode {
-                    fulfillCompletion(.success(data))
-                } else {
-                    fulfillCompletion(.failure(NetworkError.httpStatusCode(statusCode)))
-                }
-            } else if let error = error {
-                fulfillCompletion(.failure(NetworkError.urlRequestError(error)))
-            } else {
-                fulfillCompletion(.failure(NetworkError.urlSessionError))
-            }
-        })
-        task.resume()
-        return task
-} }
